Second Commit
Aspect fill changes made to the post and about screens, along with
adding the app icons to the project. Text box for initial posting text
made larger through the source code but still unable to allow the text
to layer properly, nor will it yet start at the top of the box.